Relocate getItems function to the RepositoryLayer class
