+++ ModelsSpec/ModelsSpec.swift
@@ -13,11 +13,5 @@ class ItemSpec: QuickSpec {
                 expect(brie.name).to(equal("Brie"))
             }
         }
-        
-        describe("fetch items") {
-            it("returns a list of items") {
-                expect(Item.getItems()).to(equal(Item.testData))
-            }
-        }
     }
 }
